+++ Classes/Bootstrap/TCA/TtContent.php
@@ -1,0 +1,44 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the TYPO3 extension: easyshop_package.
+ *
+ * It is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License, either version 2
+ * of the License, or any later version.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ *
+ * The TYPO3 project - inspiring people to share!
+ */
+
+namespace Wacon\Easyshop\Bootstrap\TCA;
+
+use TYPO3\CMS\Extbase\Utility\ExtensionUtility;
+use Wacon\Easyshop\Bootstrap\Base;
+
+class TtContent extends Base
+{
+    /**
+     * Does the main class purpose
+     */
+    public function invoke()
+    {
+        $this->registerPlugins();
+    }
+
+    /**
+     * Register new frontend plugins
+     */
+    private function registerPlugins()
+    {
+        ExtensionUtility::registerPlugin(
+            $this->getExtensionKeyAsNamespace(),
+            'ListWithDetail',
+            $this->getLLL('locallang_plugins.xlf:listwithdetail'),
+        );
+    }
+}