+++ CHANGES.md
@@ -1,3 +1,7 @@
+## Version 0.2.0
+- [WIP] Create frontend plugin for list view
+- [NOTICE] Update runTests.sh to work with newer GitHub Action version
+
 ## Version 0.1.0
 - [FEATURE] Define product model in TCA